Refactor types into crates
