+++ Cargo.toml
@@ -4,5 +4,4 @@ members = [
     "mvn-assembler",
     "mvn-linker",
     "mvn-utils",
-    "mvn-types",
 ]
