+++ mvn-assembler/src/types/label.rs
