+++ mvn-assembler/src/types/mneumonic/normal.rs
