+++ mvn-linker/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod parser;
 pub mod processor;
+pub mod types;
 mod writer;
