+++ mvn-linker/src/parser/mod.rs
@@ -6,12 +6,12 @@ pub mod program;
 
 pub use utils::error;
 
-pub type Position = u32;
+use crate::types::AddressPosition;
 
 pub trait Parse<'a>: Sized {
     fn parse_machine_code(input: error::Span<'a>) -> error::LocatedIResult<'a, Self>;
 }
 
 pub trait Relocate: Sized {
-    fn relocate(self, base: Position) -> Self;
+    fn relocate(self, base: AddressPosition) -> Self;
 }
