+++ mvn-linker/src/parser/program.rs
@@ -2,52 +2,13 @@ use nom::character::complete::line_ending;
 use nom::combinator::map;
 use nom::multi::separated_list1;
 
-use types::{Instruction, mneumonic::RelationalMneumonic};
+use crate::types::{AddressedLine, AddressedProgram};
 
-use super::{Parse, Position, Relocate};
+use super::{Parse, AddressPosition, Relocate};
 use super::error;
-use super::line::AddressedLine;
-
-type Lines<'a> = Vec<AddressedLine<'a>>;
-
-#[derive(Debug)]
-pub struct AddressedProgram<'a> {
-    pub lines: Lines<'a>,
-}
-
-impl<'a> AddressedProgram<'a> {
-    pub fn new(lines: Lines<'a>) -> Self {
-        Self { lines }
-    }
-
-    pub fn partition(self) -> (Lines<'a>, Lines<'a>, Lines<'a>) {
-        let (symbol_table, instructions): (Vec<AddressedLine>, Vec<AddressedLine>) = self.lines.into_iter().partition(
-            |line| line.relational_annotation.is_some()
-        );
-        let (imports, exports): (Vec<AddressedLine>, Vec<AddressedLine>) = symbol_table.into_iter().partition(
-            |line| {
-                match &line.relational_annotation {
-                    Some(line) => match &line.operation.instruction.value {
-                        Instruction::Relational(mneumonic) => mneumonic == &RelationalMneumonic::Import,
-                        _ => false,
-                    }
-                    None => false,
-                }
-            }
-        );
-        (imports, exports, instructions)
-    }
-
-    pub fn get_last_position(&self) -> Position {
-        self.lines.iter().max_by_key(
-            |line| line.address.position
-        )
-        .map_or(0, |line| line.address.position)
-    }
-}
 
 impl Relocate for AddressedProgram<'_> {
-    fn relocate(self, base: Position) -> Self {
+    fn relocate(self, base: AddressPosition) -> Self {
         AddressedProgram::new(self.lines.into_iter().map(|line| line.relocate(base)).collect())
     }
 }
@@ -61,31 +22,13 @@ impl<'a> Parse<'a> for AddressedProgram<'a> {
     }
 }
 
-impl<'a> IntoIterator for AddressedProgram<'a> {
-    type Item = AddressedLine<'a>;
-    type IntoIter = std::vec::IntoIter<Self::Item>;
-
-    fn into_iter(self) -> Self::IntoIter {
-        self.lines.into_iter()
-    }
-}
-
-impl<'a> FromIterator<AddressedLine<'a>> for AddressedProgram<'a> {
-    fn from_iter<T: IntoIterator<Item = AddressedLine<'a>>>(iter: T) -> Self {
-        let mut lines: Lines = Vec::new();
-        for line in iter {
-            lines.push(line);
-        }
-        Self::new(lines)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use pretty_assertions::assert_eq;
-    use types::{Line, Operation, Token, Position};
+    use assembler::types::Line;
+    use utils::types::*;
+    use crate::types::*;
     use super::*;
-    use crate::parser::line::AddressedLine;
 
     #[test]
     fn should_parse_program() {
