--- mvn-types/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "mvn-types"
-version = "0.1.0-alpha"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-dotenv_codegen = "0.15.0"
-utils = { path = "../mvn-utils", package = "mvn-utils" }
-
-[dev-dependencies]
-pretty_assertions = "1.2.1"
